+++ src/queue.ts
@@ -4,10 +4,12 @@ export type Worker = () => Promise<any>
 
 export type QueueOptions = {
   concurrency?: number
+  autostart?: boolean
 }
 
 export interface IQueue {
-  length: number
+  waiting: number
+  running: number
   push: (worker: Worker) => void
   start: () => void
   drain: () => Promise<void>
@@ -18,17 +20,23 @@ class Queue extends QueueEventEmitter implements IQueue {
   private isProcessing = false
   private numRunning = 0
   private concurrency: number
+  private autostart: boolean
 
   constructor(options: QueueOptions = {}) {
     super()
 
-    const { concurrency = 4 } = options
+    const { concurrency = 4, autostart = false } = options
     this.concurrency = concurrency
+    this.autostart = autostart
   }
 
 
   push(worker: Worker) {
     this.workers.push(worker)
+
+    if (this.autostart) {
+      this.process()
+    }
   }
 
   start() {
@@ -45,8 +53,12 @@ class Queue extends QueueEventEmitter implements IQueue {
     return this.eventMethod<void>('drain')()
   }
 
-  get length() {
+  get waiting() {
     return this.workers.length
+  }
+
+  get running() {
+    return this.numRunning
   }
 
   private process() {